backend/requirements.txt
fastapi
uvicorn
google-genai
qdrant-client
python-dotenv
asyncpg
pydantic
pydantic-settings
python-jose[cryptography]
bcrypt
httpx
better-auth
authlib
